Added New Command Tests And refactor'd Test locations
